+++ transceiver/_resources.py
@@ -1,4 +1,4 @@
 __build__ = '1.0'
-__version__ = '1.0.16'
+__version__ = '1.0.17'
 __author__ = 'Deskent'
 __appname__ = "Nirax Transceiver"
