+++ transceiver/src/views/api_transceiver.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, status
 
 from config import settings, logger
-from src.classes.requesters import MainRequester
+from src.classes.main_requester import MainRequester
 from src.schemas.schemas import InputSchema, OutputSchema
 
 router = APIRouter(prefix='/transceiver', tags=['Transceiver'], include_in_schema=settings.DEBUG)
